Moved github based edge case test into separate test class
